Extract tamper server from server component
Closes #9
